Extract handleResource + related utils
This is a bit heavier refactoring than I like to do
in one go. getSerialization() function signature now
includes passing in rdfaTypes. Small smell...but we
can fix that later.
